+++ st2actions/st2actions/query/base.py
@@ -28,9 +28,15 @@ from st2common.persistence.executionstate import ActionExecutionState
 from st2common.persistence.liveaction import LiveAction
 from st2common.services import executions
 from st2common.util.action_db import (get_action_by_ref, get_runnertype_by_name)
+from st2common.util import date as date_utils
 
 LOG = logging.getLogger(__name__)
 DONE_STATES = [LIVEACTION_STATUS_FAILED, LIVEACTION_STATUS_SUCCEEDED]
+
+__all__ = [
+    'Querier',
+    'QueryContext'
+]
 
 
 @six.add_metaclass(abc.ABCMeta)
@@ -119,12 +125,20 @@ class Querier(object):
         liveaction_db = LiveAction.get_by_id(execution_id)
         if not liveaction_db:
             raise Exception('No DB model for liveaction_id: %s' % execution_id)
+
         liveaction_db.result = results
         liveaction_db.status = status
+
+        done = status in DONE_STATES
+        if done and not liveaction_db.end_timestamp:
+            # Action has completed, record end_timestamp
+            liveaction_db.end_timestamp = date_utils.get_datetime_utc_now()
+
         # update liveaction, update actionexecution and then publish update.
         updated_liveaction = LiveAction.add_or_update(liveaction_db, publish=False)
         executions.update_execution(updated_liveaction)
         LiveAction.publish_update(updated_liveaction)
+
         return updated_liveaction
 
     def _invoke_post_run(self, actionexec_db, action_db):