+++ st2common/st2common/triggers.py
@@ -13,17 +13,45 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import six
+
+from mongoengine import ValidationError
 from oslo_config import cfg
 
 from st2common import log as logging
 from st2common.constants.triggers import (INTERNAL_TRIGGER_TYPES, ACTION_SENSOR_TRIGGER)
-from st2common.services.triggers import create_trigger_type_db
+from st2common.exceptions.db import StackStormDBObjectConflictError
+from st2common.services.triggers import create_trigger_type_db, create_shadow_trigger
 
 __all__ = [
     'register_internal_trigger_types'
 ]
 
 LOG = logging.getLogger(__name__)
+
+
+def _register_internal_trigger_type(trigger_definition):
+    trigger_type_db = create_trigger_type_db(trigger_type=trigger_definition)
+
+    # trigger types with parameters do no require a shadow trigger.
+    if not trigger_type_db.parameters_schema:
+        LOG.info('Registered trigger: %s.', trigger_definition['name'])
+        try:
+            trigger_db = create_shadow_trigger(trigger_type_db)
+
+            extra = {'trigger_db': trigger_db}
+            LOG.audit('Trigger created for parameter-less TriggerType. Trigger.id=%s' %
+                      (trigger_db.id), extra=extra)
+        except (ValidationError, ValueError):
+            LOG.exception('Validation failed in shadow trigger. TriggerType=%s.',
+                          trigger_type_db.get_reference().ref)
+            raise
+        except StackStormDBObjectConflictError:
+            LOG.exception('Shadow trigger creation of "%s" failed with uniqueness conflict.',
+                          trigger_type_db.get_reference().ref)
+            raise
+
+    return trigger_type_db
 
 
 def register_internal_trigger_types():
@@ -34,7 +62,9 @@ def register_internal_trigger_types():
     """
     action_sensor_enabled = cfg.CONF.action_sensor.enable
 
-    for resource_name, trigger_definitions in INTERNAL_TRIGGER_TYPES.items():
+    registered_trigger_types_db = []
+
+    for _, trigger_definitions in six.iteritems(INTERNAL_TRIGGER_TYPES):
         for trigger_definition in trigger_definitions:
             LOG.debug('Registering internal trigger: %s', trigger_definition['name'])
 
@@ -42,5 +72,7 @@ def register_internal_trigger_types():
             if is_action_trigger and not action_sensor_enabled:
                 continue
 
-            create_trigger_type_db(trigger_type=trigger_definition)
-            LOG.info('Registered trigger: %s.', trigger_definition['name'])
+            trigger_type_db = _register_internal_trigger_type(trigger_definition=trigger_definition)
+            registered_trigger_types_db.append(trigger_type_db)
+
+    return registered_trigger_types_db